feat(deps): migrate to gpui-kit 0.6.0 (gpui-pre) with form+query overrides

+++ src/features/pages/query_devtools_v2/helpers.rs
@@ -72,7 +72,7 @@ pub(super) fn filter_button(
 // Helpers
 // ---------------------------------------------------------------------------
 
-pub(super) fn format_cache_age(age_ms: Option<u128>) -> String {
+pub(super) fn format_cache_age(age_ms: Option<u64>) -> String {
     match age_ms {
         None => "n/a".to_string(),
         Some(ms) => {
